fix: refactor for code-review and fix date format bug
